fix(diet): dish-log refactor

+++ diet-service/src/main/kotlin/com/rmp/diet/dto/dish/DishDto.kt
@@ -12,5 +12,5 @@ data class DishDto (
     val fat: Double,
     val carbohydrates: Double,
     val timeToCook: Int,
-    val typeId: Int,
+    val typeId: Long,
 )
